feature/GPMS-2 - 1) Added Dockerfile 2) Fixed role mapping bug 3) Refactored TestController with NPE checks 4) Updated application properties 5) Standardized package naming
