+++ src/main/java/com/gameplatform/mainservice/security/MdcCorrelationFilter.java
@@ -1,4 +1,4 @@
-package com.gameplatform.main_service.security;
+package com.gameplatform.mainservice.security;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
