+++ src/main/java/com/gameplatform/mainservice/security/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.gameplatform.main_service.security;
+package com.gameplatform.mainservice.security;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
